requirements.txt
androguard
boto3
cachetools
celery
coloredlogs
dnspython
elasticsearch
geocoder
geoip2
ioc-finder
iocextract
lief
netaddr
pdfminer.six
pefile
piexif
pip-chill
pre-commit
pytest
python-magic
python-slugify
python-whois
rarfile
scapy
ssdeep
tld
tox
validators
virustotal-api
yara-python
